Model FIX
เปลี่ยน type ของเงิน (amount,rate,total) จาก INT เป็น float
เพิ่ม listapprove (boolean) ใน  requestlist

+++ model/loan.js
@@ -6,9 +6,9 @@ module.exports = (sequelize, type) => {
           autoIncrement: true
         },
         status : type.STRING,
-        amount: type.INTEGER,
-        interestrate : type.INTEGER,
-        totalamount : type.INTEGER,
+        amount: type.FLOAT,
+        interestrate : type.FLOAT,
+        totalamount : type.FLOAT,
         time : type.DATE
     },{
         tableName: 'loan',
