+++ model/loanlist.js
@@ -5,9 +5,9 @@ module.exports = (sequelize, type) => {
           primaryKey: true,
           autoIncrement: true
         },
-        amount: type.INTEGER,
+        amount: type.FLOAT,
         DATE : type.DATE,
-        loanleft : type.INTEGER
+        loanleft : type.FLOAT
 
     },{
         tableName: 'loanlist',
